Hoping this fixes the issue..

+++ src/index.ts
@@ -2,7 +2,7 @@ import * as fs from 'fs-extra';
 import * as ts from 'typescript';
 import * as tstl from 'typescript-to-lua';
 
-import { existsSync, writeFileSync } from 'fs';
+import { existsSync, mkdirSync, writeFileSync } from 'fs';
 import { PipeWrenchConfig, PipeWrenchConfigSchema } from './config';
 
 import Ajv from 'ajv';
@@ -359,10 +359,21 @@ class PipeWrenchPlugin implements tstl.Plugin {
           return `${key}=${value}`;
         }
       );
-      writeFileSync(
-        path.join(modSubDir, 'media/lua/shared/pipewrench_fixes.lua'),
-        PIPEWRENCH_FIXES
-      );
+      if (!existsSync(path.join(modSubDir, 'media'))) {
+        mkdirSync(path.join(modSubDir, 'media'));
+      }
+      if (!existsSync(path.join(modSubDir, 'media/lua'))) {
+        mkdirSync(path.join(modSubDir, 'media/lua'));
+      }
+      if (!existsSync(path.join(modSubDir, 'media/lua/shared'))) {
+        mkdirSync(path.join(modSubDir, 'media/lua/shared'));
+      }
+      if (!existsSync(path.join(modSubDir, 'media/lua/shared/pipewrench_fixes.lua'))) {
+        writeFileSync(
+          path.join(modSubDir, 'media/lua/shared/pipewrench_fixes.lua'),
+          PIPEWRENCH_FIXES
+        );
+      }
       writeFileSync(path.join(modSubDir, 'mod.info'), modInfoArray.join('\n'));
       result.map((file) => {
         const { outDir } = options;
